skills/news_crawl/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
